缺陷[#22080]:【组织区域】组织机构删除父节点时，子节点未同步删除，导致重复机构ID无法再次录入
    [描述说明]:
    [问题原因]: 后端逻辑问题
    [解决方法]:
    [测试建议]:

+++ iot-center-platform/src/main/java/com/yjh/platform/module/user/service/SysOrgService.java
@@ -40,7 +40,12 @@ public class SysOrgService{
     @Logs(title = "删除", code = "module", content = "删除组织机构")
     @Transactional(rollbackFor = Exception.class)
     public int deleteByPrimaryId(Long orgId) {
-        return this.sysOrgDao.deleteByPrimaryId(orgId);
+        List<Long> list = sysOrgDao.selectDownId(orgId);
+        if(list != null && list.size() > 0){
+           return sysOrgDao.batchDelete(list);
+        }else {
+            return this.sysOrgDao.deleteByPrimaryId(orgId);
+        }
     }
 
     @Logs(title = "更新", code = "module", content = "更新组织机构")
